refactor(application): Improve update logic (features/application-services)
The WorkoutProgramService previously managed the state of the
WorkoutProgram aggregate by manually removing and re-adding child
WorkoutExercisePlan entities during an update.

This commit refactors the service to delegate this responsibility to the
WorkoutProgram domain entity itself. A new `Update` method on the
aggregate root now encapsulates the logic for updating its properties
and managing its collection of exercises.

Additional improvements include:
- Using LINQ for more concise creation of exercise plans.
- Enforcing consistent use of curly braces for all control statements.
- Enhancing readability by replacing a ternary throw with an if block.